+++ src/models/components/toolparameterdefault.ts
@@ -14,11 +14,17 @@ import {
 } from "./toolparametertransform.js";
 
 /**
- * The default value for a parameter of the tool call.
+ * The default value for a parameter of a tool call.
  */
 export type ToolParameterDefault = {
   /**
-   * A transform to be applied to the value of a parameter.
+   * A transform to be applied to the value of a tool parameter.
+   *
+   * @remarks
+   *
+   * Either `value` or `format` must be set:
+   * - `value` is any arbitrary value: string, list or dictionary.
+   * - `format` is a string composed of other parameters or context variables.
    */
   transform: ToolParameterTransform;
 };