+++ src/sdk/sdk.ts
@@ -7,13 +7,15 @@ import { Agents } from "./agents.js";
 import { Channels } from "./channels.js";
 import { Conversations } from "./conversations.js";
 import { CustomMessages } from "./custommessages.js";
-import { Directory } from "./directory.js";
+import { Dashboards } from "./dashboards.js";
+import { DataSources } from "./datasources.js";
 import { Events } from "./events.js";
 import { Prompts } from "./prompts.js";
 import { Services } from "./services.js";
 import { SessionLabels } from "./sessionlabels.js";
 import { Sessions } from "./sessions.js";
 import { Tools } from "./tools.js";
+import { V1 } from "./v1.js";
 
 export class SyllableSDK extends ClientSDK {
   private _agents?: Agents;
@@ -29,6 +31,11 @@ export class SyllableSDK extends ClientSDK {
   private _conversations?: Conversations;
   get conversations(): Conversations {
     return (this._conversations ??= new Conversations(this._options));
+  }
+
+  private _dataSources?: DataSources;
+  get dataSources(): DataSources {
+    return (this._dataSources ??= new DataSources(this._options));
   }
 
   private _events?: Events;
@@ -66,8 +73,13 @@ export class SyllableSDK extends ClientSDK {
     return (this._tools ??= new Tools(this._options));
   }
 
-  private _directory?: Directory;
-  get directory(): Directory {
-    return (this._directory ??= new Directory(this._options));
+  private _dashboards?: Dashboards;
+  get dashboards(): Dashboards {
+    return (this._dashboards ??= new Dashboards(this._options));
+  }
+
+  private _v1?: V1;
+  get v1(): V1 {
+    return (this._v1 ??= new V1(this._options));
   }
 }